+++ CHANGES.md
@@ -40,6 +40,7 @@ Changes in 12.5
 - JS remove inline code from program in MassCreateAssignments.php & MassCreateAssignments.js
 - JS remove inline code from StudentAssignmentSubmissionOutput() in StudentAssignments.fnc.php & StudentAssignmentSubmissionOutput.js
 - JS remove inline code from program in ChooseCourse.php & ChooseCourse.js
+- JS remove inline code from program in ChooseRequest.php & ChooseRequest.js
 
 Changes in 12.4.3
 -----------------